refactor: move sampler and remover components into separate file with rate limiters
